backend/requirements.txt
# FastAPI 核心框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# 数据库相关
sqlalchemy>=2.0.23
# sqlite3 是 Python 标准库，无需安装

# 身份验证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP 客户端
httpx>=0.25.2
requests>=2.31.0

# AI 模型调用
openai>=1.0.0

# 数据验证
pydantic>=2.5.0
email-validator>=2.1.0

# 环境变量管理
python-dotenv>=1.0.0

# 日期时间处理
python-dateutil>=2.8.2

# CORS 支持
# 注意：FastAPI 内置 CORS 支持，无需额外安装 fastapi-cors

# 开发工具
pytest>=7.4.3
pytest-asyncio>=0.21.1